fix(pop-node): remove duplicated code
