+++ runtime/common/src/lib.rs
@@ -78,3 +78,60 @@ pub const UNINCLUDED_SEGMENT_CAPACITY: u32 = 1;
 /// How many parachain blocks are processed by the relay chain per parent. Limits the
 /// number of blocks authored per slot.
 pub const BLOCK_PROCESSING_VELOCITY: u32 = 1;
+
+/// Proxy commons for Pop runtimes
+pub mod proxy {
+
+	use super::*;
+	use codec::{Decode, Encode, MaxEncodedLen};
+	use frame_support::parameter_types;
+	use sp_runtime::RuntimeDebug;
+
+	parameter_types! {
+		// One storage item; key size 32, value size 8; .
+		pub const ProxyDepositBase: Balance = deposit(1, 40);
+		// Additional storage item size of 33 bytes.
+		pub const ProxyDepositFactor: Balance = deposit(0, 33);
+		pub const MaxProxies: u16 = 32;
+		// One storage item; key size 32, value size 16
+		pub const AnnouncementDepositBase: Balance = deposit(1, 48);
+		pub const AnnouncementDepositFactor: Balance = deposit(0, 66);
+		pub const MaxPending: u16 = 32;
+	}
+
+	/// The type used to represent the kinds of proxying allowed.
+	#[derive(
+		Copy,
+		Clone,
+		Eq,
+		PartialEq,
+		Ord,
+		PartialOrd,
+		Encode,
+		Decode,
+		RuntimeDebug,
+		MaxEncodedLen,
+		scale_info::TypeInfo,
+	)]
+	pub enum ProxyType {
+		/// Fully permissioned proxy. Can execute any call on behalf of _proxied_.
+		Any,
+		/// Can execute any call that does not transfer funds or assets.
+		NonTransfer,
+		/// Proxy with the ability to reject time-delay proxy announcements.
+		CancelProxy,
+		/// Assets proxy. Can execute any call from `assets`, **including asset transfers**.
+		Assets,
+		/// Owner proxy. Can execute calls related to asset ownership.
+		AssetOwner,
+		/// Asset manager. Can execute calls related to asset management.
+		AssetManager,
+		/// Collator selection proxy. Can execute calls related to collator selection mechanism.
+		Collator,
+	}
+	impl Default for ProxyType {
+		fn default() -> Self {
+			Self::Any
+		}
+	}
+}
